robotlib/MultiClient/requirements.txt
2ping==4.3
alembic==1.4.3
aniso8601==8.0.0
Appium-Python-Client==1.0.2
APScheduler==3.5.1
argcomplete==1.8.1
argparse-manpage==1.1
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.2.0
bleach==2.1.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cnocr==1.2.2
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
distro==1.3.0
docutils==0.16
dogtail==0.9.9
entrypoints==0.3
Flask==1.0.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
GDAL==2.4.0
gluoncv==0.6.0
graphviz==0.8.4
html5lib==1.1
idna==2.6
IPy==0.83
itsdangerous==0.24
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
kitchen==1.2.6
kiwisolver==1.3.0
ldtp==3.5.1
lib389==1.4.0.1
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.3.2
more-itertools==4.2.0
MouseInfo==0.1.3
mxnet==1.6.0
networkx==2.2
numpy==1.19.3
onboard==1.4.1
pexpect==4.6.0
Pillow==5.4.1
pluggy==0.8.0
portalocker==2.0.0
py==1.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.52
pycairo==1.16.2
pycrypto==2.6.1
PyGetWindow==0.0.9
PyGObject==3.30.4
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pysmbc==1.0.15.6
pytesseract==0.3.6
pytest==3.10.1
python-dateutil==2.7.3
python-editor==1.0.4
python-ldap==3.1.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.4
pyxdg==0.25
PyYAML==3.13
reportlab==3.5.13
requests==2.18.4
robotframework==3.0.3
robotframework-appiumlibrary==1.4.6
robotframework-requests==0.4.7
robotframework-seleniumlibrary==3.1.1
scipy==1.5.3
SecretStorage==2.3.1
selenium==3.141.0
sepolicy==1.1
setools==4.2.0
six==1.12.0
SQLAlchemy==1.2.7
tqdm==4.51.0
tzlocal==2.1
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
